integrated snapshot ring buffer

+++ bridge/contracts/interfaces/ISnapshots.sol
@@ -44,11 +44,11 @@ interface ISnapshots {
 
     function getEpochLength() external view returns (uint256);
 
-    function getChainIdFromSnapshot(uint256 epoch_) external view returns (uint256);
+    function getChainIdFromSnapshot(uint32 epoch_) external view returns (uint256);
 
     function getChainIdFromLatestSnapshot() external view returns (uint256);
 
-    function getBlockClaimsFromSnapshot(uint256 epoch_)
+    function getBlockClaimsFromSnapshot(uint32 epoch_)
         external
         view
         returns (BClaimsParserLibrary.BClaims memory);
@@ -58,20 +58,20 @@ interface ISnapshots {
         view
         returns (BClaimsParserLibrary.BClaims memory);
 
-    function getCommittedHeightFromSnapshot(uint256 epoch_) external view returns (uint256);
+    function getCommittedHeightFromSnapshot(uint32 epoch_) external view returns (uint256);
 
     function getCommittedHeightFromLatestSnapshot() external view returns (uint256);
 
-    function getAliceNetHeightFromSnapshot(uint256 epoch_) external view returns (uint256);
+    function getAliceNetHeightFromSnapshot(uint32 epoch_) external view returns (uint256);
 
     function getAliceNetHeightFromLatestSnapshot() external view returns (uint256);
 
-    function getSnapshot(uint256 epoch_) external view returns (Snapshot memory);
+    function getSnapshot(uint32 epoch_) external view returns (Snapshot memory);
 
     function getLatestSnapshot() external view returns (Snapshot memory);
 
     function getEpochFromHeight(uint256 height) external view returns (uint256);
-
+    
     function mayValidatorSnapshot(
         uint256 numValidators,
         uint256 myIdx,
